requirements.txt
requests>=2.31.0,<3
python-dotenv>=1.0.1,<2
feedparser>=6.0.10,<7
google-auth-oauthlib>=1.2.0,<2
google-api-python-client>=2.0.0,<3

